Moved github webhook into its own file github.py
