+++ zproject/urls.py
@@ -155,7 +155,7 @@ urlpatterns += patterns('zerver.views',
 
     # These are integration-specific web hook callbacks
     url(r'^api/v1/external/beanstalk$',     'webhooks.api_beanstalk_webhook'),
-    url(r'^api/v1/external/github$',        'webhooks.api_github_landing'),
+    url(r'^api/v1/external/github$',        'webhooks.github.api_github_landing'),
     url(r'^api/v1/external/jira$',          'webhooks.api_jira_webhook'),
     url(r'^api/v1/external/pivotal$',       'webhooks.api_pivotal_webhook'),
     url(r'^api/v1/external/newrelic$',      'webhooks.api_newrelic_webhook'),
